Add digest streams suite; migrate existing DigestStream tests
