+++ src/tests/streams/digest/BUILD.bazel
@@ -1,0 +1,18 @@
+load("//:build/wd_test.bzl", "wd_test")
+
+# The digest streams suite runs the same test modules under two configs:
+# digest-cpp against the C++ implementation, digest-ts against the
+# TypeScript implementation.
+
+digest_suite_srcs = glob(["*.js"]) + ["digest-modules.capnp"]
+
+wd_test(
+    src = "digest-cpp.wd-test",
+    data = digest_suite_srcs,
+)
+
+wd_test(
+    src = "digest-ts.wd-test",
+    args = ["--experimental"],
+    data = digest_suite_srcs,
+)
