+++ src/workerd/api/tests/BUILD.bazel
@@ -514,18 +514,6 @@ wd_test(
 )
 
 wd_test(
-    src = "crypto-streams-test.wd-test",
-    args = ["--experimental"],
-    data = ["crypto-streams-test.js"],
-)
-
-wd_test(
-    src = "crypto-streams-ts-test.wd-test",
-    args = ["--experimental"],
-    data = ["crypto-streams-test.js"],
-)
-
-wd_test(
     src = "data-url-fetch-test.wd-test",
     args = ["--experimental"],
     data = ["data-url-fetch-test.js"],
@@ -1377,12 +1365,6 @@ wd_test(
 wd_test(
     src = "htmlrewriter-transform-cancel-test.wd-test",
     data = ["htmlrewriter-transform-cancel-test.js"],
-)
-
-wd_test(
-    src = "ts-digest-stream-test.wd-test",
-    args = ["--experimental"],
-    data = ["ts-digest-stream-test.js"],
 )
 
 wd_test(
