--- src/workerd/api/tests/crypto-streams-test.wd-test
@@ -1,14 +1,0 @@
-using Workerd = import "/workerd/workerd.capnp";
-
-const unitTests :Workerd.Config = (
-  services = [
-    ( name = "crypto-streams-test",
-      worker = (
-        modules = [
-          (name = "worker", esModule = embed "crypto-streams-test.js")
-        ],
-        compatibilityFlags = ["nodejs_compat"],
-      )
-    ),
-  ],
-);
