--- src/workerd/api/tests/ts-digest-stream-test.wd-test
@@ -1,21 +1,0 @@
-using Workerd = import "/workerd/workerd.capnp";
-
-const unitTests :Workerd.Config = (
-  services = [
-    ( name = "ts-digest-stream-test",
-      worker = (
-        modules = [
-          (name = "worker", esModule = embed "ts-digest-stream-test.js")
-        ],
-        compatibilityFlags = [
-          "nodejs_compat",
-          "typescript_implemented_streams",
-          "experimental",
-        ],
-      )
-    ),
-  ],
-  autogates = [
-    "workerd-autogate-per-isolate-javascript-bootstrap",
-  ],
-);
